+++ base/widgets/ActiveForm.php
@@ -63,31 +63,12 @@ class ActiveForm extends \yii\widgets\ActiveForm
         //Html::addCssClass($config, "sx-input-int");
         $config['class'] = ArrayHelper::getValue($config, 'class') . "form-control sx-input-int";
 
-
-        /**
-         *  'initval' => 3.00,
-            'min' => 0,
-            'max' => 100,
-            'step' => 0.1,
-            'decimals' => 2,
-            'boostat' => 5,
-            'maxboostedstep' => 10,
-            'prefix' => '$',
-         */
-
-        $defaultOptions = [
-            'pluginOptions' =>
-            [
-                'max' => 9999999999
-            ]
+        $defaultConfig = [
+            'type' => 'number'
         ];
 
-        $pluginOptions = ArrayHelper::merge($defaultOptions, (array) $config);
-
-        //return $this->field($model, $attribute, $fieldOptions)->textInput($config);
-        return $this->field($model, $attribute, $fieldOptions)->widget(
-            \kartik\touchspin\TouchSpin::className(), $pluginOptions
-        );
+        $config = ArrayHelper::merge($defaultConfig, (array) $config);
+        return $this->field($model, $attribute, $fieldOptions)->textInput($config);
     }
 
     /**